Move migration API to correct location
