+++ simplyblock_web/api/v2/cluster/storage_pool/volume/__init__.py
@@ -9,9 +9,10 @@ from simplyblock_core import utils as core_utils
 from simplyblock_core.controllers import backup_controller, lvol_controller, snapshot_controller
 from simplyblock_core.models.lvol_model import LVol
 
-from ..._dependencies import Cluster, StoragePool, Volume
-from ..._dtos import BackupDTO, VolumeDTO, SnapshotDTO, TaskDTO
-from ... import util
+from ...._dependencies import Cluster, StoragePool, Volume
+from ...._dtos import BackupDTO, VolumeDTO, SnapshotDTO, TaskDTO
+from .... import util
+from .migration import api as migration_api
 
 
 api = APIRouter()
@@ -368,4 +369,5 @@ def delete_backups(cluster: Cluster, pool: StoragePool, volume: Volume) -> Respo
     return Response(status_code=204)
 
 
+instance_api.include_router(migration_api, prefix="/migrations")
 api.include_router(instance_api)
