+++ simplyblock_web/api/v2/cluster/storage_pool/volume/migration.py
@@ -10,14 +10,14 @@ from simplyblock_core.db_controller import DBController
 from simplyblock_core.exceptions import MigrationConflictError, PreconditionError
 from simplyblock_web import utils
 
-from .._dependencies import Cluster, Migration
-from .._dtos import MigrationDTO
-from ..util import CreationResponseFormatParameter, creation_response
+from ...._dependencies import Cluster, Migration
+from ...._dtos import MigrationDTO
+from ....util import CreationResponseFormatParameter, creation_response
 
 api = APIRouter()
 
 
-@api.get('/', name='clusters:migrations:list')
+@api.get('/', name='clusters:storage-pools:volumes:migrations:list')
 def list_migrations(cluster: Cluster) -> List[MigrationDTO]:
     db = DBController()
     migrations = db.get_migrations(cluster.get_id())
@@ -31,7 +31,7 @@ class _PreCreateParams(BaseModel):
     host_nqn: Annotated[str, Field(pattern=utils.NQN_PATTERN)] | None = None
 
 
-@api.post('/', name='clusters:migrations:create', status_code=201, responses={201: {"content": None}})
+@api.post('/', name='cluster:storage-pools:volumes:migrations:create', status_code=201, responses={201: {"content": None}})
 def create_migration(request: Request, cluster: Cluster, parameters: _PreCreateParams, response_format: CreationResponseFormatParameter = "identifier") -> Response:
     try:
         migration_id, connect_strings = migration_controller.create_migration(
@@ -46,7 +46,7 @@ def create_migration(request: Request, cluster: Cluster, parameters: _PreCreateP
     return creation_response(
         request, response_format,
         entity_id=UUID(migration_id),
-        route_name='clusters:migrations:detail',
+        route_name='cluster:storage-pools:volumes:migrations:detail',
         route_kwargs={'cluster_id': UUID(cluster.get_id()), 'migration_id': UUID(migration_id)},
         get_full=lambda id: MigrationDTO.from_model(
             db.get_migration_by_id(str(id)), connect_strings=connect_strings),
@@ -56,7 +56,7 @@ def create_migration(request: Request, cluster: Cluster, parameters: _PreCreateP
 instance_api = APIRouter(prefix='/{migration_id}')
 
 
-@instance_api.get('/', name='clusters:migrations:detail')
+@instance_api.get('/', name='cluster:storage-pools:volumes:migrations:detail')
 def get_migration(cluster: Cluster, migration: Migration) -> MigrationDTO:
     return MigrationDTO.from_model(migration)
 
@@ -66,7 +66,7 @@ class _ContinueParams(BaseModel):
     deadline_seconds: int = 14400
 
 
-@instance_api.post('/continue', name='clusters:migrations:continue', status_code=200)
+@instance_api.post('/continue', name='cluster:storage-pools:volumes:migrations:continue', status_code=200)
 def continue_migration(cluster: Cluster, migration: Migration, parameters: _ContinueParams):
     try:
         migration_id = migration_controller.start_migration(
@@ -79,7 +79,7 @@ def continue_migration(cluster: Cluster, migration: Migration, parameters: _Cont
     return {"migration_id": migration_id}
 
 
-@instance_api.delete('/', name='clusters:migrations:cancel', status_code=200)
+@instance_api.delete('/', name='cluster:storage-pools:volumes:migrations:cancel', status_code=200)
 def cancel_migration(cluster: Cluster, migration: Migration):
     ok, error = migration_controller.cancel_migration(migration.get_id())
     if not ok:
